# docs/openapi/mvp.yaml
openapi: 3.0.3
info:
  title: StrollBar MVP API
  version: 0.1.0
  description: REST contract for the StrollBar MVP.
servers:
  - url: https://api.strollbar.local/v1
...
paths:
...
  /strolls/{strollId}/stages:
    get:
      tags: [Stages]
      summary: List ordered stages for a stroll
      parameters:
        - $ref: '#/components/parameters/StrollId'
      responses:
        '200':
          description: Ordered stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stage'

    post:
      tags: [Stages]
      summary: Add stage to stroll
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StrollId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStageRequest'
      responses:
        '201':
          description: Stage created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stage'
  /strolls/{strollId}/stages/reorder:
    patch:
      tags: [Stages]
      summary: Reorder stroll stages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StrollId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderStagesRequest'
      responses:
        '200':
          description: Stages reordered
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StrollId:
      in: path
      name: strollId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Stage:
      type: object
      required:
        - id
        - strollId
        - orderIndex
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
        strollId:
          type: string
          format: uuid
        orderIndex:
          type: integer
          minimum: 1
        name:
          type: string
        description:
          type: string
        notes:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        videoUrls:
          type: array
          items:
            type: string
            format: uri
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
          minimum: -90
          maximum: 90
        longitude:
          type: number
          nullable: true
          minimum: -180
          maximum: 180
...
    CreateStageRequest:
      type: object
      required: [orderIndex, name, description]
      properties:
        orderIndex:
          type: integer
          minimum: 1
        name:
          type: string
        description:
          type: string
        notes:
          type: string
        imageUrls:
          type: array
          items:
            type: string
            format: uri
        videoUrls:
          type: array
          items:
            type: string
            format: uri
        address:
          type: string
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        riddleAnswer:
          type: string
    ReorderStagesRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [stageId, orderIndex]
            properties:
              stageId:
                type: string
                format: uuid
              orderIndex:
                type: integer
                minimum: 1
...